+++ pcpicker_webserver/nbproject/jax-ws.xml
@@ -4,6 +4,9 @@
     <service name="Pcpicker_webservice">
       <implementation-class>pcpicker.Pcpicker_webservice</implementation-class>
     </service>
+    <service name="Pcpicker_webserviceForDesktop">
+      <implementation-class>Pcpicker_webserviceForDesktop.Pcpicker_webserviceForDesktop</implementation-class>
+    </service>
   </services>
   <clients/>
 </jax-ws>
